Created a working version (read notes below)
The correct command to run this iteration of the system requires you to
force a build of the docker images, because the base image doesn't exist
in any remote repository (and never will). After the first build, you
don't have to repeat this (but it wouldn't hurt):

  `docker compose up --build`

Do not specify the `-p` option to change the project name or the
children images will not be able to find that local base image (the name
will change).